ekf2: split vel_pos_fusion.cpp
