+++ src/modules/ekf2/EKF/position_fusion.cpp
@@ -1,6 +1,6 @@
 /****************************************************************************
  *
- *   Copyright (c) 2015-2023 PX4 Development Team. All rights reserved.
+ *   Copyright (c) 2015-2024 PX4 Development Team. All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
@@ -31,63 +31,7 @@
  *
  ****************************************************************************/
 
-/**
- * @file vel_pos_fusion.cpp
- *
- * @author Roman Bast <bapstroman@gmail.com>
- * @author Siddharth Bharat Purohit <siddharthbharatpurohit@gmail.com>
- * @author Paul Riseborough <p_riseborough@live.com.au>
- *
- */
-
-#include <mathlib/mathlib.h>
 #include "ekf.h"
-
-void Ekf::updateVelocityAidSrcStatus(const uint64_t &time_us, const Vector2f &obs, const Vector2f &obs_var,
-				     const float innov_gate, estimator_aid_source2d_s &aid_src) const
-{
-	resetEstimatorAidStatus(aid_src);
-
-	for (int i = 0; i < 2; i++) {
-		aid_src.observation[i] = obs(i);
-		aid_src.innovation[i] = _state.vel(i) - aid_src.observation[i];
-
-		aid_src.observation_variance[i] = math::max(sq(0.01f), obs_var(i));
-		const int state_index = State::vel.idx + i;
-		aid_src.innovation_variance[i] = P(state_index, state_index) + aid_src.observation_variance[i];
-	}
-
-	setEstimatorAidStatusTestRatio(aid_src, innov_gate);
-
-	aid_src.timestamp_sample = time_us;
-}
-
-void Ekf::updateVelocityAidSrcStatus(const uint64_t &time_us, const Vector3f &obs, const Vector3f &obs_var,
-				     const float innov_gate, estimator_aid_source3d_s &aid_src) const
-{
-	resetEstimatorAidStatus(aid_src);
-
-	for (int i = 0; i < 3; i++) {
-		aid_src.observation[i] = obs(i);
-		aid_src.innovation[i] = _state.vel(i) - aid_src.observation[i];
-
-		aid_src.observation_variance[i] = math::max(sq(0.01f), obs_var(i));
-		const int state_index = State::vel.idx + i;
-		aid_src.innovation_variance[i] = P(state_index, state_index) + aid_src.observation_variance[i];
-	}
-
-	setEstimatorAidStatusTestRatio(aid_src, innov_gate);
-
-	// vz special case if there is bad vertical acceleration data, then don't reject measurement,
-	// but limit innovation to prevent spikes that could destabilise the filter
-	if (_fault_status.flags.bad_acc_vertical && aid_src.innovation_rejected) {
-		const float innov_limit = innov_gate * sqrtf(aid_src.innovation_variance[2]);
-		aid_src.innovation[2] = math::constrain(aid_src.innovation[2], -innov_limit, innov_limit);
-		aid_src.innovation_rejected = false;
-	}
-
-	aid_src.timestamp_sample = time_us;
-}
 
 void Ekf::updateVerticalPositionAidSrcStatus(const uint64_t &time_us, const float obs, const float obs_var,
 		const float innov_gate, estimator_aid_source1d_s &aid_src) const
@@ -130,44 +74,6 @@ void Ekf::updateHorizontalPositionAidSrcStatus(const uint64_t &time_us, const Ve
 	setEstimatorAidStatusTestRatio(aid_src, innov_gate);
 
 	aid_src.timestamp_sample = time_us;
-}
-
-void Ekf::fuseVelocity(estimator_aid_source2d_s &aid_src)
-{
-	if (!aid_src.innovation_rejected) {
-		// vx, vy
-		if (fuseDirectStateMeasurement(aid_src.innovation[0], aid_src.innovation_variance[0], State::vel.idx)
-		    && fuseDirectStateMeasurement(aid_src.innovation[1], aid_src.innovation_variance[1], State::vel.idx + 1)
-		   ) {
-			aid_src.fused = true;
-			aid_src.time_last_fuse = _time_delayed_us;
-
-			_time_last_hor_vel_fuse = _time_delayed_us;
-
-		} else {
-			aid_src.fused = false;
-		}
-	}
-}
-
-void Ekf::fuseVelocity(estimator_aid_source3d_s &aid_src)
-{
-	if (!aid_src.innovation_rejected) {
-		// vx, vy, vz
-		if (fuseDirectStateMeasurement(aid_src.innovation[0], aid_src.innovation_variance[0], State::vel.idx + 0)
-		    && fuseDirectStateMeasurement(aid_src.innovation[1], aid_src.innovation_variance[1], State::vel.idx + 1)
-		    && fuseDirectStateMeasurement(aid_src.innovation[2], aid_src.innovation_variance[2], State::vel.idx + 2)
-		   ) {
-			aid_src.fused = true;
-			aid_src.time_last_fuse = _time_delayed_us;
-
-			_time_last_hor_vel_fuse = _time_delayed_us;
-			_time_last_ver_vel_fuse = _time_delayed_us;
-
-		} else {
-			aid_src.fused = false;
-		}
-	}
 }
 
 void Ekf::fuseHorizontalPosition(estimator_aid_source2d_s &aid_src)
